@@ -3,9 +3,10 @@
   "version": "1.0.0",
   "description": "Facebook management suite",
   "scripts": {
-    "dev": "webpack -w --progress",
-    "build": "webpack --progress",
-    "deploy": "NODE_ENV=production webpack --progress && ./script/deploy",
+    "dev": "webpack -w --progress --config webpack.dev.js",
+    "build": "webpack --progress --config webpack.dev.js",
+    "build-prod": "NODE_ENV=production webpack --progress --config webpack.prod.js",
+    "deploy": "NODE_ENV=production webpack --progress --config webpack.prod.js && ./script/deploy",
     "prestart": "npm i",
     "start": "node server.js"
   },
@@ -13,7 +14,10 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.9.1",
+    "bootstrap": "^3.3.7",
     "form-data": "^2.3.1",
+    "jquery": "^2.2.1",
+    "muicss": "^0.9.27",
     "react": "^0.14.7",
     "react-bootstrap": "^0.31.2",
     "react-bs-notifier": "^4.3.2",
@@ -22,31 +26,27 @@
     "react-router-dom": "^4.2.2",
     "socket.io-client": "^2.0.3",
     "store": "^2.0.12",
-    "universal-cookie": "^2.0.8",
     "url-search-params-polyfill": "^2.0.1",
     "uuid": "^3.1.0"
   },
   "devDependencies": {
+    "webpack": "^3.7.1",
+    "sass-loader": "^3.1.2",
+    "json-loader": "^0.5.7",
     "babel-core": "^6.5.1",
-    "babel-loader": "^6.2.2",
+    "babel-loader": "^7.1.1",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
-    "bootstrap": "^3.3.7",
+    "node-sass": "^3.4.2",
     "css-loader": "^0.23.1",
     "express": "^4.13.4",
     "extract-loader": "^1.0.1",
     "file-loader": "^0.11.2",
     "html-loader": "^0.5.1",
-    "jquery": "^2.2.1",
-    "json-loader": "^0.5.7",
-    "muicss": "^0.9.24",
-    "node-sass": "^3.4.2",
-    "react-spinners": "0.0.30",
-    "sass-loader": "^3.1.2",
     "script-loader": "^0.6.1",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.9",
-    "webpack": "^1.12.13"
+    "webpack-merge": "^4.1.0"
   }
 }
